# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Granja Salvadoreña",
        "description": "API REST para el sistema de E-commerce de Granja Salvadoreña",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Obtener historial de órdenes",
                "description": "Obtiene las órdenes del usuario autenticado.",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Órdenes obtenidas correctamente."
                    },
                    "401": {
                        "description": "No autenticado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Crear una orden",
                "description": "Crea una nueva orden para el usuario autenticado.",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada correctamente."
                    },
                    "401": {
                        "description": "No autenticado."
                    },
                    "422": {
                        "description": "Datos de la orden inválidos."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Consultar una orden",
                "description": "Obtiene una orden perteneciente al usuario autenticado.",
                "operationId": "8114ac185cc1bf301d8862617e497098",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden obtenida correctamente."
                    },
                    "403": {
                        "description": "La orden no pertenece al usuario."
                    },
                    "404": {
                        "description": "Orden no encontrada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}/cancel": {
            "patch": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Cancelar una orden",
                "description": "Cancela una orden perteneciente al usuario autenticado.",
                "operationId": "0b59b76911446d0e1563044ad919d7b9",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden cancelada correctamente."
                    },
                    "403": {
                        "description": "La orden no pertenece al usuario."
                    },
                    "422": {
                        "description": "La orden ya está cancelada."
                    },
                    "404": {
                        "description": "Orden no encontrada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}/payment": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Iniciar pago de una orden",
                "description": "Crea un PaymentIntent de Stripe y registra el pago asociado a la orden.",
                "operationId": "890b2ad463d65dc7d1b14f1a73aafab1",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pago iniciado correctamente."
                    },
                    "403": {
                        "description": "La orden no pertenece al usuario."
                    },
                    "422": {
                        "description": "La orden no puede ser pagada."
                    },
                    "500": {
                        "description": "Error al procesar el pago con Stripe."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}